requirements.txt
python-dotenv
openai==0.28.1
Flask
pytest
